rename: `icon-check` -> `complete` 파일명 변경
